+++ src/jeppeto/main.py
@@ -1,129 +1,54 @@
+#! /usr/bin/env python
+# -*- coding: UTF8 -*-
 # This file is part of  program Jeppeto
 # Copyright © 2022  Carlo Oliveira <carlo@nce.ufrj.br>,
 # `Labase <http://labase.selfip.org/>`__; `GPL <http://is.gd/3Udt>`__.
 # SPDX-License-Identifier: (GPLv3-or-later AND LGPL-2.0-only) WITH bison-exception
-from _spy.vitollino.main import Cena, Elemento, Texto, Sala, Labirinto
-from _spy.vitollino.main import INVENTARIO as inv
+"""Jeppeto - Editor Gráfico de Jogos WEB.
+
+    Classes neste módulo:
+        - :py:class:`SvgPainter`    Controla a geração de Componentes no Canvas.
+        - :py:class:`SvgMarquee`    Controla o cursor de desenho.
+        - :py:class:`Main`          Ponto de entrada central do aplicativo.
+        - :py:class:`ToolBox`       Define as ferramentas de pintura e edição.
+
+.. codeauthor:: Carlo Oliveira <carlo@nce.ufrj.br>
+
+Changelog
+---------
+
+.. versionadded::    22.10
+        First version ported from Supygirls.
+
+"""
+
 from browser import svg, document, html, alert
 from collections import namedtuple
-
-# import collections as col
-# create employee NamedTuple
-Boxer = namedtuple('Boxer', ['f', 'x', 'y', 'w', 'h'])
-NOBOX = Boxer(0, 0, 0, 100, 60)
-
-CDD = "https://upload.wikimedia.org/wikipedia/commons/e/e9/Cidade_de_Deus.jpg"
-FLASH = "https://pngriver.com/wp-content/uploads/2018/03/Download-Flash-PNG-Pic-For-Designing-Projects.png"
-# AWESOME = "https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css"
-"https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@6.1.2/css/fontawesome.min.css"
+from jeppeto.wrapper import ModelMake, Box
+# "https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@6.1.2/css/fontawesome.min.css"
 AWESOME = "https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.2/css/all.min.css"
 FAW = html.LINK(rel="stylesheet", href=AWESOME, Type="text/css")
 FAW.setAttribute("type", "text/css")
 JEPPETO = "https://i.imgur.com/eI50beC.png"
-document.head <= FAW
-
-
-class Box:
-    BOX = []
-
-    def __init__(self, box=NOBOX):
-        # Box.BOX = []
-        self.box = box
-
-    def paint(self, f=None, **kwargs):
-        box = Box(Boxer(f=f, **kwargs))
-        Box.BOX.append(box)
-
-    def remove(self, box):
-        Box.BOX.remove(box)
-
-    def as_dict(self):
-        b = self.box
-        return dict(f=b.f, x=b.x, y=b.y, w=b.w, h=b.h)
-
-    def find(self, x, y):
-        for bbox in Box.BOX:
-            box = bbox.box
-            if (box.x < x < box.x + box.w) and (box.y < y < box.y + box.h):
-                return bbox
-        return None
-
-
-class Tomada(Cena, Box):
-    def __init__(self, img='', box=NOBOX):
-        Cena.__init__(self, img)
-        Box.__init__(self, box)
-
-
-class Ator(Elemento, Box):
-    def __init__(self, img='', cena='', box=NOBOX):
-        Elemento.__init__(self, img, cena=cena)
-        Box.__init__(self, box)
-
-
-class Objeto(Elemento, Box):
-    def __init__(self, img='', cena='', box=NOBOX):
-        Elemento.__init__(self, img, cena=cena)
-        Box.__init__(self, box)
-
-
-class Fala(Texto, Box):
-    def __init__(self, cena='', fala='', box=NOBOX):
-        Texto.__init__(self, cena, fala)
-        Box.__init__(self, box)
-
-
-class Quarto(Sala, Box):
-    def __init__(self, img='', box=NOBOX):
-        Sala.__init__(self, img)
-        Box.__init__(self, box)
-
-
-class ModelMake:
-    def __init__(self, gui):
-        self.gui = gui
-        self.parts = dict(tomada=self.tomada, ator=self.ator, objeto=self.objeto, texto=self.texto, sala=self.sala)
-
-    def paint(self, box=NOBOX, **kwargs):
-        if box.f not in self.parts:
-            return
-        box = self.parts[box.f](box=box, **kwargs)
-        Box.BOX.append(box)
-
-    def tomada(self, img='', box=NOBOX):
-        return Tomada(img=img, box=box)
-
-    def ator(self, img='', box=NOBOX):
-        return Ator(img=img, box=box)
-
-    def objeto(self, img='', box=NOBOX):
-        return Objeto(img=img, box=box)
-
-    def texto(self, cena='', box=NOBOX):
-        return Fala(cena=cena, box=box)
-
-    def sala(self, img='', box=NOBOX):
-        return Quarto(img=img, box=box)
-
-
+_ = document.head <= FAW
 CW, CH = 1200, 650
 Z = 5
 PX, PY = 20, 20
 
 
 class SvgPainter:
-    def __init__(self, main):
-        self._main = main
+    def __init__(self, main_):
+        self._main = main_
         self.shape = dict(b=lambda x, y, w, h, it=self: svg.rect(
             x=x, y=y, width=w, height=h, fill=it.fill, fill_opacity=it.opacity))
         self.root = root = document["pydiv"]
         root.html = ""
         self.canvas = svg.svg(viewBox=f"{PX} {PY} {CW / Z} {CH / Z}", width=1200, height=650)
-        root <= self.canvas
+        _ = root <= self.canvas
         self.fill = "white"
         self.opacity = 1
-        self.screen = self.paint("b", x=0, y=0, w=1200, h=650)
-        Box.BOX = []
+        self.screen = self.paint("b", x=30, y=0, w=1200, h=650)
+        # Box.BOX = []
         self.fill = "red"
         self.opacity = 0.5
 
@@ -131,7 +56,7 @@ class SvgPainter:
         self.fill = f if f else self.fill
         self._main.paint(f=f, **kwargs)
         shp = self.shape[shape](**kwargs)
-        self.canvas <= shp
+        _ = self.canvas <= shp
         return shp
         # shp.setAttribute("fill-opacity", self.opacity)
 
@@ -141,9 +66,9 @@ class SvgPainter:
 
 
 class SvgMarquee:
-    def __init__(self, main, canvas, stroke="grey", dash="4 1"):
+    def __init__(self, main_, canvas, stroke="grey", dash="4 1"):
         self.zoom = Z
-        self._main = main
+        self._main = main_
         self.px, self.py = PX, PY
         self.canvas = canvas
         self.origin, self.size = (0, 0), (0, 0)
@@ -152,7 +77,7 @@ class SvgMarquee:
         rect = self.canvas.canvas.getBoundingClientRect()
         self.ox, self.oy = rect.left, rect.top
         self.vbox = self.canvas.canvas.viewBox
-        self.fill = None
+        self.fill = self.marquee = None
         self.stroke, self.dash = stroke, dash
         self.opacity = 0.8
         self.shape = lambda x, y, w, h, it=self: svg.rect(x=x, y=y, width=w, height=h, fill=it.fill,
@@ -171,7 +96,7 @@ class SvgMarquee:
         self.fill = f if f else self.fill
         shp = self.shape(**kwargs)
         self.marquee = shp
-        self.canvas.canvas <= shp
+        _ = self.canvas.canvas <= shp
         print(str(kwargs), shp)
         return shp
         # shp.setAttribute("fill-opacity", self.opacity)
@@ -214,7 +139,7 @@ class SvgMarquee:
         self.marquee = self.paint(x=0, y=0, w=0, h=0)
         if w < 2 or h < 2:
             color = ev.target.getAttribute("fill")
-            box = Boxer(f=color, x=x, y=y, w=w, h=h)
+            _ = Boxer(f=color, x=x, y=y, w=w, h=h)
             self.canvas.filler(color=color)
             self._main.select(x=dx, y=dy)
         else:
@@ -225,6 +150,7 @@ class SvgMarquee:
 class Main:
     def __init__(self, marker=None, painter=None, tool=None, make=None):
         self.model = make
+        self.filling = None
         self.painter = painter
         self.marquee = marker
         self.tool = tool
@@ -235,7 +161,7 @@ class Main:
             'background-size': 'cover', 'background-repeat': 'no-repeat',
             'min-height': "500px", 'width': "400px", 'background-image': f'url({JEPPETO})'})
         self.splash.onclick = self.main
-        self.root <= self.splash
+        _ = self.root <= self.splash
 
     def main(self, _=0):
         self.root.html = ""
@@ -252,7 +178,8 @@ class Main:
 
     def tooler(self, ev=0, tool=0):
         self.tool.html = ""
-        self.tool <= html.SPAN(Class=tool, style={'font-size': '30px', 'color': 'black'})
+        _ = ev
+        _ = self.tool <= html.SPAN(Class=tool, style={'font-size': '30px', 'color': 'black'})
 
     def paint(self, f=None, **kwargs):
         pass
@@ -263,6 +190,7 @@ class Main:
 
     def select(self, f=None, x=-1, y=-1, **kwargs):
         box = self.model.find(x, y)
+        _ = f in kwargs
         bbox = box.as_dict() if box else None
         # print(box, bbox,">>>", [(b.box.x, b.box.y) for b in Box.BOX])
         # self.marquee.paint(x=x, y=y, w=40, h=40) if box
@@ -278,19 +206,19 @@ class ToolBox:
         self.painter = painter
         self.app = app
         self.tool_action = {}
-        self.colors = self.tools = None
+        self.colors = self.tools = self.filling = self.menu = self.tool = self.marquee = None
 
     def main(self, _=0):
         self.painter = self.painter or SvgPainter(self)
         # self.menu = html.DIV(style={'position':"absolute", 'left':'10px', 'top':'100px'}) #, 'z-index': 10})
-        self.menu = html.DIV(style={'position': "absolute", 'left': '-120px', 'top': '-30px'})  # , 'z-index': 10})
-        self.app.root <= self.menu
+        self.menu = html.DIV(style={'position': "absolute", 'left': '10px', 'top': '10px'})  # , 'z-index': 10})
+        _ = self.app.root <= self.menu
 
         def sty(tp, bg):
             return {'position': "absolute", 'left': '0px', 'top': f'{tp}px',
                     'min-height': "30px", 'width': "30px", 'background-color': bg}
 
-        def lay(ev, cor):
+        def lay(ev, *_):
             # cor = ev.target.style.backgroundColor
             cor = ev.target.title
             # cor = "orange"
@@ -300,11 +228,11 @@ class ToolBox:
         colors = "transparent yellow green orange red blue cyan magenta purple".split()
         tools = ["transparent"] * 4
         ncol = len(colors) + 1
-        style = {}
+        # style = {}
         off = 40 * ncol + 40
         self.colors = [html.DIV(title=bg, style=sty(tp, bg)) for tp, bg in zip(list(range(80, ncol * 40, 40)), colors)]
         self.tools = [html.DIV(style=sty(tp, bg)) for tp, bg in zip(list(range(off, off + 4 * 40, 40)), tools)]
-        # document <= self.menu
+        _ = document <= self.menu
         edit = "fa-solid fa-paintbrush"
         select = "fa-solid fa-object-group"
         zoom = "fa-solid fa-magnifying-glass"
@@ -315,39 +243,40 @@ class ToolBox:
         texto = "fa-solid fa-comment"
         coisa = "fa-solid fa-screwdriver-wrench"
         sala = "fa-solid fa-building"
-        labirinto = "fa-solid fa-castle"
+        # labirinto = "fa-solid fa-castle"
         puzzle = "fa-solid fa-puzzle-piece"
         self.filling = html.SPAN(Class="fa-solid fa-fill", style={'font-size': '30px', 'color': 'white'})
         # tool = html.SPAN(Class=edit, style={'font-size':'30px', 'color':'black'})
         self.tool = self.colors[0]
-        self.menu <= self.filling
-        self.menu <= self.tool
+        _ = self.menu <= self.filling
+        _ = self.menu <= self.tool
         self.tooler(tool=edit)
         for col, tol_ in zip(self.colors, (cena, cena, people, roteiro, texto, coisa, sala, puzzle)):
-            self.menu <= col
-            col <= html.SPAN(Class=tol_,
-                             style={'font-size': '30px', 'pointer-events': 'none', 'color': 'black', 'opacity': 0.4})
+            _ = self.menu <= col
+            _ = col <= html.SPAN(Class=tol_,
+                                 style={'font-size': '30px', 'pointer-events': 'none', 'color': 'black',
+                                        'opacity': 0.4})
             col.bind("click", lambda ev, color=col: lay(ev, color))
         # off = 40*ncol - 40
         actions = (self.tool_edit, self.tool_select, self.tool_zoom, self.tool_turnoff)
         for _col, _tool, _act in zip(self.tools, (edit, select, zoom, turnoff), actions):
             _col.style.top = f"{off}px"
             off += 40
-            self.menu <= _col
-            _col <= html.SPAN(Class=_tool, style={'font-size': '30px', 'color': 'black'})
+            _ = self.menu <= _col
+            _ = _col <= html.SPAN(Class=_tool, style={'font-size': '30px', 'color': 'black'})
             _col.bind("click", lambda ev, tol=_tool, act=_act, it=self: it.tooler(ev=ev, tool=tol, action=act))
 
     def filler(self, color):
         self.filling.style.color = color
 
-    def tooler(self, ev=0, tool=0, action=lambda *_: None):
+    def tooler(self, ev=0, tool='', action=lambda *_: None):
         self.tool.html = ""
-        self.tool <= html.SPAN(Class=tool, style={'font-size': '30px', 'color': 'black'})
+        _ = self.tool <= html.SPAN(Class=tool, style={'font-size': '30px', 'color': 'black'})
         action(ev)
 
     def paint(self, f=None, **kwargs):
         pass
-        #self.model.paint(f=f, **kwargs)
+        # self.model.paint(f=f, **kwargs)
 
     def remove(self, box):
         self.model.remove(box)
@@ -361,10 +290,11 @@ class ToolBox:
     def tool_zoom(self, ev):
         pass
 
-    def tool_turnoff(self, ev):
+    def tool_turnoff(self, *_):
         self.app.root.html = ""
 
     def select(self, f=None, x=-1, y=-1, **kwargs):
+        _ = f in kwargs
         box = self.model.find(x, y)
         bbox = box.as_dict() if box else None
         # print(box, bbox,">>>", [(b.box.x, b.box.y) for b in Box.BOX])